machine: move disable_sentry function
